Removed VS compiler warnings (missing type cast).
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1441 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/widget.cpp
@@ -475,7 +475,7 @@ bool Widget::createRect(int radius)
                 }
                 else
                 {
-                    vertex_ya = m_height;
+                    vertex_ya = (float)m_height;
                 }
 
                 if( m_round_corners & WGT_AREA_SW )
@@ -512,7 +512,7 @@ bool Widget::createRect(int radius)
                 }
                 else
                 {
-                    vertex_ya = m_height;
+                    vertex_ya = (float)m_height;
                 }
 
                 if( m_round_corners & WGT_AREA_SE )
